% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_exact.R
\name{cc_exact}
\alias{cc_exact}
\alias{cc_exact_gauss_2Cl1D}
\alias{cc_exact_gauss_2Cl2D}
\alias{cc_exact_gauss_3Cl1D}
\alias{cc_exact_gauss_3Cl3D}
\alias{cc_exact_gauss_kCl1D}
\alias{cc_exact_gauss_2ClmD}
\alias{cc_exact_gauss_kClmD}
\alias{cc_exact_exp_2Cl}
\alias{cc_exact_exp_3Cl}
\alias{cc_exact_exp_kCl}
\alias{cc_exact_lnorm_kCl}
\alias{cc_exact_gamma_kCl}
\title{Functions for calculating channel capacity}
\usage{
cc_exact(type = "gauss", classes = 2, dimension = 1, ...)

cc_exact_gauss_2Cl1D(par, x_length, s_length)

cc_exact_gauss_2Cl2D(par, x_length, s_length)

cc_exact_gauss_3Cl1D(par, x_length, s_length)

cc_exact_gauss_3Cl3D(par, x_length, s_length)

cc_exact_gauss_kCl1D(par, x_length, mc_length, mc_cores = 1)

cc_exact_gauss_2ClmD(par, s_length, mc2_length, mc_cores = 1)

cc_exact_gauss_kClmD(par, x_length, mc_length, mc2_length, mc_cores = 1,
  popt_sugg = NULL)

cc_exact_exp_2Cl(par, x_length, s_length)

cc_exact_exp_3Cl(par, x_length, s_length)

cc_exact_exp_kCl(par, x_length, mc_length, mc_cores = 1)

cc_exact_lnorm_kCl(par, x_length, mc_length, mc_cores = 1)

cc_exact_gamma_kCl(par, x_length, mc_length, mc_cores = 1)
}
\description{
Calculating channel capacity by brute-force optimisation and numerical integration
}
\examples{
contact authors to use these functions
}
