% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_curve_main.R
\name{information_curve_main}
\alias{information_curve_main}
\title{Function for calculating Information response Curve}
\usage{
information_curve_main(dataRaw, signal = "signal", response = "response",
  side_variables = NULL, cc_maxit = 100, lr_maxit = 1000,
  MaxNWts = 5000, formula_string = NULL, output_path = NULL,
  plot_height = 4, plot_width = 6, cores = 2, scale = TRUE)
}
\arguments{
\item{dataRaw}{must be a data.frame object}

\item{signal}{is a character object that indicates columns of data to be treated as channel's input}

\item{response}{is a character vector that indicates columns of data to be treated as channel's output}

\item{side_variables}{is an optional character vector that indicates side variables' columns of data, if NULL no side variables are included}

\item{cc_maxit}{is the number of iteration of iterative algorithm of finding maximum CC}

\item{lr_maxit}{is the number of iteration of iterative algorithm of logistic regression}

\item{formula_string}{is a character object that includes a formula syntax to use in algorithm for capacity calculation}

\item{output_path}{is the directory in which output will be saved}

\item{plot_height, }{plot_width - the basic dimnesions of plots}

\item{cores}{- number of cores to use in parallel computing if implemented (only if testing=TRUE)}

\item{scale}{is the logical indicating if the data preprocessing should be carried out}

\item{maxNWts}{is the maximum acceptable number of weights in logistic regression algorithm - the higher, the slower the computations}

\item{model_out}{is the logical indicating if the calculated logisitc regression model should be included in output list}
}
\value{
a list with two elements:
\itemize{
\item output$signal - signal
\item output$cc     - channel capacity in bits
}
}
\description{
This function calculate the integral of function
}
\examples{

}
\keyword{internal}
