% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointProb_gaussian_md.R
\name{jointProb_gaussian_md}
\alias{jointProb_gaussian_md}
\title{Density function of gaussian mixture distribution}
\usage{
jointProb_gaussian_md(x, s, mu, sigma)
}
\arguments{
\item{x}{is a numerical vector of quantiles}

\item{s}{is a numerical vector of the number of gaussian distribution in the mixture}

\item{mu}{is a list, of length the same as vector s, whose each element is the mean vector of subsequent gaussians}

\item{sigma}{is a list, of length the same as vector s, whose each element is the covariance matrix of subsequent gaussians}
}
\value{
the value of density
}
\description{
This function calculates the value of density of a finite mixutre of multivariate gaussian distributions
}
\examples{
jointProb_gaussian_md(x=c(1,1),s=c(0.2,0.8), 
mu=list(c(0,0),c(1,1)),
sigma=list(matrix(c(1,0.5,0.5,1),2,2),matrix(c(1,0.1,0.1,1),2,2) ))


x=seq(from=-10,to=10,length.out = 500)
dmg=matrix(0,length(x),length(x))
for (i in 1:length(x) ){
    for (j in 1:length(x)){
        dmg[i,j]=jointProb_gaussian_md(x=c(x[i],x[j]),s=c(0.2,0.8), mu=list(c(0,0),c(1,1)),sigma=list(matrix(c(1,0.5,0.5,1),2,2),matrix(c(1,0.1,0.1,1),2,2) ))
    }
}
trapz_simple(x,apply(dmg,1,function(z) trapz_simple(x,z) ))

}
\keyword{internal}
